back-end/requirements.txt
click==8.0.3
flask==2.1.3
flask-sqlalchemy==2.5.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
requests
flask_cors
pillow
Werkzeug==2.0.2